requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
h5py==3.14.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.9
lightning==2.5.4
lightning-utilities==0.15.2
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
propcache==0.3.2
protobuf==6.32.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.4
pytz==2025.2
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
scipy==1.16.1
sentry-sdk==2.35.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
torchmetrics==1.8.1
tqdm==4.67.1
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.3
yarl==1.20.1
